requirements-dev.txt
# Main dependencies
typer>=0.19.2
pandas>=2.2.2
fastapi>=0.95.0,<1.0.0
uvicorn>=0.29.0
networkx>=3.3
matplotlib>=3.8
plotly>=5.22.0
requests>=2.32.3
requests-mock
PyYAML>=6.0.1
cryptography>=42.0.8
filelock>=3.15.4
pydantic>=2.2.2
pydantic-settings>=1.0.5

# Extras: kafka
confluent-kafka>=2.12.0
six>=1.16.0

# Extras: aws
boto3>=1.34.131

# Extras: gcs
google-cloud-storage>=2.16.0

# Extras: datadog
datadog-api-client>=2.19.0
datadog>=0.44

# Extras: dev
pytest>=8.2.2
pytest-cov>=5.0.0
pytest-mock>=3.10.0
black>=24.4.2
mypy>=1.10.1
isort>=5.13.2
coverage>=7.5.3
moto>=5.0.4
httpx>=0.28.1
loguru>=0.7.0